Move CLI code out of __main__.py to allow multiprocessing to work for 'python -m sphinxlint'
