+++ pyproject.toml
@@ -35,7 +35,7 @@ Repository = "https://github.com/sphinx-contrib/sphinx-lint"
 Changelog = "https://github.com/sphinx-contrib/sphinx-lint/releases"
 
 [project.scripts]
-sphinx-lint = "sphinxlint.__main__:main"
+sphinx-lint = "sphinxlint.cli:main"
 
 [tool.hatch]
 version.source = "vcs"
