+++ tests/test_sphinxlint.py
@@ -4,7 +4,7 @@ from sphinxlint.utils import paragraphs
 
 import pytest
 
-from sphinxlint.__main__ import main
+from sphinxlint.cli import main
 
 FIXTURE_DIR = Path(__file__).resolve().parent / "fixtures"
